+++ tokenauth/token.go
@@ -38,6 +38,7 @@ type Claims struct {
 	jwt.StandardClaims
 	OrgID         string `json:"org_id" validate:"required"`    // Organization ID
 	AppID         string `json:"app_id"`                        // Application ID
+	SessionID     string `json:"session_id"`                    // Session ID
 	Purpose       string `json:"purpose" validate:"required"`   // Token purpose (eg. access, csrf...)
 	AuthType      string `json:"auth_type" validate:"required"` // Authentication method (eg. email, phone...)
 	Permissions   string `json:"permissions"`                   // Granted permissions
@@ -48,9 +49,10 @@ type Claims struct {
 	Admin         bool   `json:"admin"`                         // Is this token for an admin?
 
 	// User Data: DO NOT USE AS IDENTIFIER OR SHARE WITH THIRD-PARTY SERVICES
-	Name  string `json:"name,omitempty"`  // User full name
-	Email string `json:"email,omitempty"` // User email address
-	Phone string `json:"phone,omitempty"` // User phone number
+	Name        string            `json:"name,omitempty"`         // User full name
+	Email       string            `json:"email,omitempty"`        // User email address
+	Phone       string            `json:"phone,omitempty"`        // User phone number
+	ExternalIDs map[string]string `json:"external_ids,omitempty"` // External user identifiers for use in external integrations
 
 	//TODO: Once the new user ID scheme has been adopted across all services these claims should be removed
 	UID string `json:"uid,omitempty"` // Unique user identifier for specified "auth_type"